program line reader: modified to carry program unit pointer
added program unit pointer member with access function
modified users of program line reader accordingly
- removed program unit argument to operand text and remove functions
- removed program unit argument from token constructor for program word
corrected type on dictionary string function index argument

+++ dictionary.h
@@ -53,7 +53,7 @@ public:
 	void clear(void);
 	uint16_t add(Token *token, EntryType *returnNewEntry = {});
 	int remove(uint16_t index);
-	std::string string(int index) const
+	std::string string(uint16_t index) const
 	{
 		return m_iterator[index]->first;
 	}
